Add files via upload
This case study is the Capstone project for the Google Data Analytics Professional Certificate. In this case study, I am a Junior Data Analyst and I am analyzing data from consumer's smart devices in order to gain insights on how they use them and look for growth opportunities for the company giving data-driven suggestions for a marketing strategy.